Add support for removing executing nodes
`schedule` calls can now optionally avoid requeueing the
currently-executing node, when there is one. This is how parking and
exiting would be implemented in a scheduler using the library.

For now, there is still no way to remove nodes that are just sitting in
the queue; they must be scheduled once and then be rescheduled without
requeuing.